% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\name{gap_diagram}
\alias{gap_diagram}
\title{Draw gap diagram}
\usage{
gap_diagram(pathways, time_end = 365)
}
\arguments{
\item{pathways}{dataframe of pathway features}

\item{time_end}{the end of the diagram}
}
\description{
Draw gap diagram
}
