% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppa.R
\name{visualise_accessibility}
\alias{visualise_accessibility}
\title{Visualise TB services: capacity, access, and referrals}
\usage{
visualise_accessibility(hs, ps, gref = NULL)
}
\arguments{
\item{hs}{dataframe of hospital features}

\item{ps}{dataframe of pathway features}

\item{gref}{figure of referral pattern}
}
\description{
Visualise TB services: capacity, access, and referrals
}
\examples{
ppa <- visualise_accessibility(pseudo.tb.h, pseudo.tb.p)
print(ppa)
}
