% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_freq.R
\name{visualise_pattern_freq}
\alias{visualise_pattern_freq}
\title{Draw pattern frequency diagram}
\usage{
visualise_pattern_freq(pathways, maps = state_space_tb$GP,
  x.lab = "Proportion of time in stages (\%)",
  y.lab = "Accumulated pathways")
}
\arguments{
\item{pathways}{json formated patient pathways}

\item{maps}{colour maps}

\item{x.lab}{label of x-axis}

\item{y.lab}{label of y-axis}
}
\description{
Draw pattern frequency diagram
}
\examples{
pat.freq <- visualise_pattern_freq(pseudo.tb.js)
print(pat.freq)
}
