% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_dist.R
\name{visualise_stage_dist}
\alias{visualise_stage_dist}
\title{Draw stage distribution diagram}
\usage{
visualise_stage_dist(pathways, maps = state_space_tb$All,
  x.lab = "Time since first visit (day)", y.lab = "Stage distribution",
  t_end = 365, dt = 1)
}
\arguments{
\item{pathways}{json formated patient pathways}

\item{maps}{colour maps}

\item{x.lab}{label of x-axis}

\item{y.lab}{label of y-axis}

\item{t_end}{read end of the diagram, default = 365 days}

\item{dt}{bar update frequency of the diagram}
}
\description{
Draw stage distribution diagram
}
\examples{
st.di <- visualise_stage_dist(pseudo.tb.js)
print(st.di)
}
